requirements.txt
addict==2.4.0
aiofiles==0.6.0
aiohttp==3.7.4
annotated-types==0.7.0
asttokens==3.0.0
async-timeout==3.0.1
attrs==25.3.0
babel==2.17.0
blis==1.2.1
catalogue==2.0.10
certifi==2025.8.3
cffi==1.17.1
chardet==3.0.4
charset-normalizer==3.4.3
click==8.1.8
cloudpathlib==0.21.1
cmd2==1.5.0
colorama==0.4.6
coloredlogs==15.0.1
commonmark==0.9.1
confection==0.1.5
cryptography==45.0.6
csvw==3.5.1
curated-tokenizers==0.0.9
curated-transformers==0.1.1
cymem==2.0.11
decorator==5.2.1
dlinfo==2.0.0
docopt==0.6.2
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl
espeakng-loader==0.2.4
exceptiongroup==1.3.0
executing==2.2.0
filelock==3.19.1
filetype==1.2.0
flatbuffers==25.2.10
fsspec==2025.7.0
ftfy==6.3.1
grpcio==1.74.0
hf-xet==1.1.8
huggingface-hub==0.34.4
humanfriendly==10.0
idna==2.10
ipython==8.18.1
isodate==0.7.2
jedi==0.19.2
Jinja2==3.0.0
joblib==1.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
kokoro==0.7.16
langcodes==3.5.0
language-tags==1.2.0
language_data==1.3.0
loguru==0.7.3
marisa-trie==1.3.0
markdown-it-py==3.0.0
marker==2.1.5
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mdurl==0.1.2
misaki==0.9.4
mpmath==1.3.0
multidict==6.6.4
murmurhash==1.0.13
networkx==3.2.1
num2words==0.5.14
numpy==1.26.4
onnxruntime==1.19.2
opencv-python==4.11.0.86
packaging==25.0
parso==0.8.4
pdfminer.six==20250506
pdfplumber==0.11.7
pdftext==0.3.14
pexpect==4.9.0
phonemizer-fork==3.3.2
pillow==10.4.0
piper-tts==1.3.0
preshed==3.0.10
prompt_toolkit==3.0.51
propcache==0.3.2
protobuf==6.32.0
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.19.2
pyparsing==3.2.3
pypdfium2==4.30.0
pyperclip==1.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
PyYAML==6.0.2
RapidFuzz==3.13.0
rdflib==7.1.4
referencing==0.36.2
regex==2024.11.6
requests==2.25.1
rfc3986==1.5.0
rich==10.2.0
rpds-py==0.27.0
ruamel.yaml==0.17.17
ruamel.yaml.clib==0.2.12
safetensors==0.6.2
scikit-learn==1.4.2
scipy==1.13.1
segments==2.3.0
shellingham==1.5.4
six==1.17.0
smart_open==7.3.0.post1
soundfile==0.13.1
spacy==3.8.7
spacy-curated-transformers==0.3.1
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
stack-data==0.6.3
surya-ocr==0.5.0
sympy==1.14.0
tabulate==0.9.0
texify==0.1.10
thinc==8.3.4
threadpoolctl==3.6.0
tokenizers==0.21.4
torch==2.8.0
tqdm==4.67.1
traitlets==5.14.3
transformers==4.55.2
typing-inspection==0.4.1
typing_extensions==4.14.1
uritemplate==4.2.0
urllib3==1.26.20
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.4.1
wordninja==2.0.0
wrapt==1.17.3
yarl==1.20.1
